% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_ellipse.R
\name{nice_ellipse}
\alias{nice_ellipse}
\title{nMDS biplot with centroid ellipses}
\usage{
nice_ellipse(out_perma_nmds, mat_ech, var_test)
}
\arguments{
\item{out_perma_nmds}{a list, the output of perma_nmds}

\item{mat_ech}{a data frame or matrix, the matrix specifying the sample groups with modalities in columns and samples in rows.}

\item{var_test}{a character, the name of the modality to test.}
}
\value{
A ggplot object
}
\description{
This function display the samples in the first 2 axis of nMDS and the centroids of each group
}
