% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_facet.R
\name{nice_facet}
\alias{nice_facet}
\title{Represent nMDS with facets}
\usage{
nice_facet(out_perma_nmds, mat_ech, var_facet, var_col)
}
\arguments{
\item{out_perma_nmds}{the output of the `perma_nmds()` function}

\item{mat_ech}{a dataframe or matrix, the matrix specifying the samples groups
with modalities in columns and samples in rows}

\item{var_facet}{a vector containing 2 variables to make a facet grid}

\item{var_col}{a group variable to color nMDS points}
}
\value{
Return an nMDS plot from the `ggplot` package
}
\description{
This function create a graphical output for a nMDS with facets
}
