% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_nmds.R
\name{perma_nmds}
\alias{perma_nmds}
\title{Execute nMDS and PERMANOVA}
\usage{
perma_nmds(mat_com, mat_ech, var_test, hellinger)
}
\arguments{
\item{mat_com}{a dataframe or matrix, the community matrix with species
in columns and samples in rows}

\item{mat_ech}{a dataframe or matrix, the matrix specifying the samples groups
with modalities in columns and samples in rows}

\item{var_test}{a vector containing variables to test}

\item{hellinger}{A boolean, if TRUE a Hellinger transformation is applied}
}
\value{
Return a list containing 2 elements, first an nMDS from the `vegan`
package, second a dataframe r2 and pvalue for each variable tested
}
\description{
This function execute a PERMANOVA test to asses
the significance of the effect of variable on community
structure. It also performs an nMDS using the Bray-Curtis dissimilarity measure.
}
