% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_nmds_ph.R
\name{perma_nmds_ph}
\alias{perma_nmds_ph}
\title{Post_hoc test for PERMANOVA}
\usage{
perma_nmds_ph(mat_com, mat_ech, var_test)
}
\arguments{
\item{mat_com}{a data frame or matrix, the community matrix with species in columns and samples in rows.}

\item{mat_ech}{a data frame or matrix, the matrix specifying the sample groups with modalities in columns and samples in rows.}

\item{var_test}{a character, the name of the modality to test.}
}
\value{
A data frame containing r2 and p values for each group couple
}
\description{
This function computes pairwise PERMANOVAs for each couple of groups
}
